requirements.txt
langchain
langchain-openai
langchain-core
openai
python-dotenv
langdetect
asyncpg
pinecone
pandas
openpyxl
psycopg2-binary
fastapi
uvicorn
gunicorn
jmespath
langgraph
langchain-community
langchainhub
tiktoken
langchain-pinecone
langchain-text-splitters 
streamlit
flask
redis==5.2.0